requirements.txt
pandas==2.1.0
numpy==1.24.0
scikit-learn==1.3.0
dvc==3.30.0
pyyaml==6.0.1
mlflow==2.9.0
fastapi==0.104.0
uvicorn==0.24.0
evidently==0.4.33
xgboost==2.0.0
lightgbm==4.1.0